(*.[ch]):	patches to types to wipe warnings from compilers 	with more strict noise levels. (also good when compiling 	with gcc -Wall -pedantic -Wwrite-strings -Wcast-qual)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2584 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/host/hr_utils.h
@@ -3,5 +3,5 @@
  *
  */
 
-u_char * date_n_time (time_t * , int * );
+u_char * date_n_time (time_t * , size_t * );
 time_t ctime_to_timet (char* );
